Rename src --> ngraph_bridge
