+++ ngraph_bridge/enable_variable_ops/ngraph_catalog.cc
