+++ ngraph_bridge/enable_variable_ops/ngraph_enter_in_catalog.h
