+++ ngraph_bridge/enable_variable_ops/ngraph_replace_op_utilities.cc
@@ -18,8 +18,8 @@
 #include "tensorflow/core/graph/node_builder.h"
 #include "tensorflow/core/graph/types.h"
 
-#include "src/ngraph_mark_for_clustering.h"
-#include "src/ngraph_utils.h"
+#include "ngraph_bridge/ngraph_mark_for_clustering.h"
+#include "ngraph_bridge/ngraph_utils.h"
 
 using namespace std;
