+++ ngraph_bridge/enable_variable_ops/ngraph_replace_variable_modifiers.cc
@@ -20,8 +20,8 @@
 #include "ngraph_capture_variables.h"
 #include "ngraph_replace_op_utilities.h"
 #include "ngraph_replace_variable_modifiers.h"
-#include "src/ngraph_api.h"
-#include "src/ngraph_utils.h"
+#include "ngraph_bridge/ngraph_api.h"
+#include "ngraph_bridge/ngraph_utils.h"
 
 using namespace std;
 namespace ng = ngraph;
