+++ ngraph_bridge/enable_variable_ops/ngraph_replace_variable_modifiers.h
