+++ ngraph_bridge/enable_variable_ops/ngraph_rewrite_pass.cc
@@ -24,14 +24,14 @@
 #include "ngraph_enter_in_catalog.h"
 #include "ngraph_replace_variable_modifiers.h"
 #include "ngraph_rewrite_for_tracking.h"
-#include "src/ngraph_api.h"
-#include "src/ngraph_assign_clusters.h"
-#include "src/ngraph_capture_variables.h"
-#include "src/ngraph_cluster_manager.h"
-#include "src/ngraph_deassign_clusters.h"
-#include "src/ngraph_encapsulate_clusters.h"
-#include "src/ngraph_mark_for_clustering.h"
-#include "src/ngraph_utils.h"
+#include "ngraph_bridge/ngraph_api.h"
+#include "ngraph_bridge/ngraph_assign_clusters.h"
+#include "ngraph_bridge/ngraph_capture_variables.h"
+#include "ngraph_bridge/ngraph_cluster_manager.h"
+#include "ngraph_bridge/ngraph_deassign_clusters.h"
+#include "ngraph_bridge/ngraph_encapsulate_clusters.h"
+#include "ngraph_bridge/ngraph_mark_for_clustering.h"
+#include "ngraph_bridge/ngraph_utils.h"
 
 #if defined NGRAPH_DISTRIBUTED
 #include "ngraph/distributed.hpp"
