+++ ngraph_bridge/enable_variable_ops/ngraph_variable_modifiers.cc
@@ -28,10 +28,10 @@
 #include "ngraph/runtime/backend.hpp"
 #include "ngraph_catalog.h"
 #include "ngraph_var.h"
-#include "src/ngraph_backend_manager.h"
-#include "src/ngraph_freshness_tracker.h"
-#include "src/ngraph_timer.h"
-#include "src/ngraph_utils.h"
+#include "ngraph_bridge/ngraph_backend_manager.h"
+#include "ngraph_bridge/ngraph_freshness_tracker.h"
+#include "ngraph_bridge/ngraph_timer.h"
+#include "ngraph_bridge/ngraph_utils.h"
 
 using namespace std;
 namespace ng = ngraph;
