+++ ngraph_bridge/grappler/ngraph_optimizer.cc
@@ -15,8 +15,8 @@
  *******************************************************************************/
 
 #include "ngraph_optimizer.h"
-#include "src/ngraph_backend_manager.h"
-#include "src/ngraph_cluster_manager.h"
+#include "ngraph_bridge/ngraph_backend_manager.h"
+#include "ngraph_bridge/ngraph_cluster_manager.h"
 
 #include "tensorflow/core/framework/attr_value.pb.h"
 #include "tensorflow/core/framework/node_def.pb.h"
