+++ ngraph_bridge/ngraph_assign_clusters.cc
