+++ ngraph_bridge/ngraph_backend_manager.cc
