+++ ngraph_bridge/ngraph_capture_variables.cc
