+++ ngraph_bridge/ngraph_cluster_manager.cc
