+++ ngraph_bridge/ngraph_encapsulate_clusters.cc
