+++ ngraph_bridge/ngraph_encapsulate_op.cc
