+++ ngraph_bridge/ngraph_rewrite_pass.cc
