+++ ngraph_bridge/ngraph_tracked_variable.cc
