+++ ngraph_bridge/ngraph_utils.cc
