+++ ngraph_bridge/tf_graphcycles.h
