+++ ngraph_bridge/version.cc
