+++ ngraph_bridge/version.h
